rename to cli_logging
