+++ slidescore_api/cli_logging.py
